refactor(build): extract companyLogoResolver from jobCardHtml
